--- .changeset/forty-lies-fry.md
@@ -1,5 +1,0 @@
----
-'@hono/next-auth': major
----
-
-initial support for next auth
