+++ .changeset/red-coins-attend.md
@@ -1,0 +1,5 @@
+---
+'@hono/auth-js': major
+---
+
+initial support auth.js with hono
