+++ package.json
@@ -25,7 +25,7 @@
     "build:esbuild-transpiler": "yarn workspace @hono/esbuild-transpiler build",
     "build:oauth-providers": "yarn workspace @hono/oauth-providers build",
     "build:react-renderer": "yarn workspace @hono/react-renderer build",
-    "build:next-auth": "yarn workspace @hono/next-auth build",
+    "build:auth-js": "yarn workspace @hono/auth-js build",
     "build": "run-p 'build:*'",
     "lint": "eslint 'packages/**/*.{ts,tsx}'",
     "lint:fix": "eslint --fix 'packages/**/*.{ts,tsx}'",
