+++ packages/auth-js/src/client.ts
