+++ packages/auth-js/src/index.ts
