+++ packages/auth-js/tsconfig.json
