+++ packages/auth-js/tsup.config.ts
