+++ packages/auth-js/vitest.config.mts
